Updated visual design to improve legibility
We have improved the legibility of this document and the visual clarity
of some of the diagrams.

The words are unchanged from the original v1.0 of the document.

Includes an updated PDF document (PDF/X-1a:2001 standard) and the
source Adobe InDesign files to allow others to change the visual design
in future.